backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
cohere>=5.0.0
qdrant-client>=1.7.0
asyncpg>=0.29.0
pydantic>=2.5.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic-settings>=2.0.0
httpx>=0.25.0